Solutions/2021/AdventOfCode2021.Tests/Day09/Day09SolverTests.cs
namespace AdventOfCode2021.Day09.Tests;

public class Day09SolverTests
{
    public static TheoryData<int, string[]> PuzzleOneTestData
        = new TheoryData<int, string[]>()
        {
                {
                    15,
                    new []
                    {
"2199943210",
"3987894921",
"9856789892",
"8767896789",
"9899965678",
                    }
                }
        };

    [Theory]
    [MemberData(nameof(PuzzleOneTestData))]
    public void PuzzleOne(int expected, params string[] input)
    {
        var actual = Day09Solver.PuzzleOne(input);

        Assert.Equal(expected, actual);
    }

    public static TheoryData<int, string[]> PuzzleTwoTestData
        = new TheoryData<int, string[]>()
        {
                {
                    1134,
                    new []
                    {
"2199943210",
"3987894921",
"9856789892",
"8767896789",
"9899965678",
                    }
                }
        };

    [Theory]
    [MemberData(nameof(PuzzleTwoTestData))]
    public void PuzzleTwo(int expected, params string[] input)
    {
        var actual = Day09Solver.PuzzleTwo(input);

        Assert.Equal(expected, actual);
    }
}